+++ bot/discord/src/jvmMain/kotlin/io/github/sophon/discord/DiscordBot.kt
@@ -105,13 +105,13 @@ internal class DiscordBotImpl(
 
         when {
             botOutput.embedBuilder != null -> {
-                createEmbedMessage(botOutput.embedBuilder)
+                createEmbedMessage(botOutput.embedBuilder, botOutput.images)
             }
             botOutput.plainText != null -> {
                 createPlainMessage(botOutput.plainText)
             }
             botOutput.errorEmbedBuilder != null -> {
-                createEmbedMessage(botOutput.errorEmbedBuilder)
+                createEmbedMessage(botOutput.errorEmbedBuilder, )
                     .delete(delay = TIME_DELETE_ERROR_EMBED, scope = kord)
             }
         }
@@ -136,7 +136,7 @@ internal class DiscordBotImpl(
 
         when {
             botOutput.embedBuilder != null -> {
-                createEmbedResponse(botOutput.embedBuilder)
+                createEmbedResponse(botOutput.embedBuilder, botOutput.images)
             }
             botOutput.plainText != null -> {
                 createPlainResponse(botOutput.plainText)