+++ core/src/commonMain/kotlin/io/github/sophon/core/wiki/domain/model/Move.kt
@@ -35,7 +35,8 @@ data class Move(
     data class Urls(
         val characterWiki: String? = null,
         val videoId: String? = null, //TODO: change to videoUrl
-        val hitboxImage: String? = null,
+        val hitboxImageList: List<String> = listOf(),
+        val wikiUrl: String,
     )
 
     @Serializable