+++ feat/wikiDustLoop/src/commonMain/kotlin/io/github/sophon/wikidustloop/data/MoveMapper.kt
@@ -5,6 +5,7 @@ import io.github.sophon.core.util.urlDecode
 import io.github.sophon.core.wiki.domain.model.Move
 
 internal fun MoveListResponseDto.toDomain(
+    gameId: String,
     imageUrlMap: Map<String, String>,
 ): List<Move> {
     return cargoQuery.map { wrapper ->
@@ -32,11 +33,11 @@ internal fun MoveListResponseDto.toDomain(
             notes = dto.notes.formNotes(),
 
             urls = Move.Urls(
-                hitboxImage = dto.hitboxes
-                    ?.split(",")
+                hitboxImageList = dto.hitboxes
                     .orEmpty()
-                    .firstOrNull()
-                    ?.let { imageUrlMap[it] }
+                    .split(";")
+                    .mapNotNull { imageUrlMap.getOrElse(key = it, defaultValue = { null }) },
+                wikiUrl = formMoveWikiUrl(gameId, dto),
             ),
 
             airDashProperties = Move.AirDashProperties(
@@ -78,4 +79,8 @@ internal fun String?.formNotes(): List<String> {
         ?.map { it.trim() }
         ?.filter { it.isNotBlank() }
         ?: emptyList()
+}
+
+internal fun formMoveWikiUrl(gameId: String, dto: MoveDto): String {
+    return "${dto.chara.formWikiUrl(gameId)}#${dto.input}"
 }