+++ CHANGELOG.md
@@ -5,9 +5,16 @@
 - core v1.1.1
   - sync with bot version
   - bot status
-- dustLoop v1.0.1
+- DreamCancel v1.0.2
+  - multiple hitbox images
+- DustLoop v1.0.1
   - clickable Unique Movement Options
   - fixed note formatting
+  - multiple hitbox images
+- SuperCombo v1.0.2
+  - clickable moves
+- Wavu v1.0.3
+  - clickable moves
 
 ## [v1.1.0] - 2025-12-08
 - core v1.0.6